add swagger and verify authorization by communicate authentication microservice

+++ src/main/java/ma/fstt/microserviceexaminateur/controllers/ExaminaterController.java
@@ -7,6 +7,7 @@ import ma.fstt.microserviceexaminateur.payload.request.UpdateExaminaterRequest;
 import ma.fstt.microserviceexaminateur.payload.response.MessageResponse;
 import ma.fstt.microserviceexaminateur.repository.ExaminaterRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
@@ -27,38 +28,56 @@ public class ExaminaterController {
     PasswordEncoder encoder;
     @Autowired
     public ExaminaterRepository examinaterRepository;
+    private final AuthService authService;
+    public ExaminaterController(ExaminaterRepository examinaterRepository, AuthService authService) {
+        this.examinaterRepository = examinaterRepository;
+        this.authService = authService;
+    }
 
     public String encodeDate(Date date) {
         SimpleDateFormat sdf = new SimpleDateFormat("MM/dd/yyyy");
         return encoder.encode(sdf.format(date));
     }
     @PostMapping("/add")
-    public ResponseEntity<MessageResponse> addExaminater(@RequestBody AddExaminaterRequest addExaminaterRequest) {
+    public ResponseEntity<MessageResponse> addExaminater(@RequestBody AddExaminaterRequest addExaminaterRequest,@RequestHeader("Authorization") String authorizationHeader) {
         try {
-            Examinater examinater = new Examinater(
-                    addExaminaterRequest.getL_name(),
-                    addExaminaterRequest.getF_name(),
-                    addExaminaterRequest.getL_name()+"_"+addExaminaterRequest.getF_name(),
-                    encodeDate(addExaminaterRequest.getDate_of_birth()),
-                    addExaminaterRequest.getEmail(),
-                    addExaminaterRequest.getPhone(),
-                    addExaminaterRequest.getCity(),
-                    addExaminaterRequest.getAddress(),
-                    addExaminaterRequest.getCin(),
-                    addExaminaterRequest.getDate_of_birth()
-            );
 
-            examinaterRepository.save(examinater);
+            // Extract the token from the Authorization header
+            String token = extractTokenFromHeader(authorizationHeader);
+                if (!authService.isValidEmployeeToken(token)) {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageResponse(401, "Not authorized"));
+                }
+                Examinater examinater = new Examinater(
+                        addExaminaterRequest.getL_name(),
+                        addExaminaterRequest.getF_name(),
+                        addExaminaterRequest.getL_name()+"_"+addExaminaterRequest.getF_name(),
+                        encodeDate(addExaminaterRequest.getDate_of_birth()),
+                        addExaminaterRequest.getEmail(),
+                        addExaminaterRequest.getPhone(),
+                        addExaminaterRequest.getCity(),
+                        addExaminaterRequest.getAddress(),
+                        addExaminaterRequest.getCin(),
+                        addExaminaterRequest.getDate_of_birth()
+                );
 
-            return ResponseEntity.ok(new MessageResponse(201,"Examinater saved successfully"));
+                examinaterRepository.save(examinater);
+
+                return ResponseEntity.ok(new MessageResponse(201,"Examinater saved successfully"));
+
+
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(new MessageResponse(400,"Examinater doesn't save "));
 
         }
     }
     @GetMapping("/getAll")
-    public ResponseEntity<List<Examinater>> getAllExaminaters() {
+    public ResponseEntity<Object> getAllExaminaters(@RequestHeader("Authorization") String authorizationHeader) {
         try {
+            // Extract the token from the Authorization header
+            String token = extractTokenFromHeader(authorizationHeader);
+            if (!authService.isValidEmployeeToken(token)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageResponse(401, "Not authorized"));
+            }
             List<Examinater> examinaters = examinaterRepository.findAll();
             List<Examinater> filteredExaminaters = examinaters.stream()
                     .filter(examinater -> hasRole(examinater, "ROLE_EXAMINATER"))
@@ -74,8 +93,13 @@ public class ExaminaterController {
     }
 
     @GetMapping("/{examinaterId}")
-    public ResponseEntity<Examinater> getExaminaterById(@PathVariable String examinaterId) {
+    public ResponseEntity<Object> getExaminaterById(@PathVariable String examinaterId,@RequestHeader("Authorization") String authorizationHeader) {
         try {
+            // Extract the token from the Authorization header
+            String token = extractTokenFromHeader(authorizationHeader);
+            if (!authService.isValidEmployeeToken(token)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageResponse(401, "Not authorized"));
+            }
             Optional<Examinater> examinater = examinaterRepository.findById(examinaterId);
             if (examinater.isPresent() &&  hasRole(examinater.get(), "ROLE_EXAMINATER")) {
                 return ResponseEntity.ok(examinater.get());
@@ -88,12 +112,22 @@ public class ExaminaterController {
     }
 
 
-
+    private String extractTokenFromHeader(String authorizationHeader) {
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            return authorizationHeader.substring(7);
+        }
+        return null;
+    }
 
 
     @DeleteMapping("/{examinaterId}")
-    public ResponseEntity<MessageResponse> deleteExaminater(@PathVariable String examinaterId) {
+    public ResponseEntity<MessageResponse> deleteExaminater(@PathVariable String examinaterId,@RequestHeader("Authorization") String authorizationHeader) {
         try {
+            // Extract the token from the Authorization header
+            String token = extractTokenFromHeader(authorizationHeader);
+            if (!authService.isValidEmployeeToken(token)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageResponse(401, "Not authorized"));
+            }
             Optional<Examinater> examinater = examinaterRepository.findById(examinaterId);
             if (examinater.isPresent()) {
                 examinaterRepository.delete(examinater.get());
@@ -111,8 +145,14 @@ public class ExaminaterController {
     public ResponseEntity<MessageResponse> examinaterUpdate(
             @PathVariable String examinaterId,
             @RequestBody UpdateExaminaterRequest updatedClientRequest
+            ,@RequestHeader("Authorization") String authorizationHeader
     ) {
         try {
+            // Extract the token from the Authorization header
+            String token = extractTokenFromHeader(authorizationHeader);
+            if (!authService.isValidEmployeeToken(token)) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new MessageResponse(401, "Not authorized"));
+            }
             Optional<Examinater> optionalExaminater = examinaterRepository.findById(examinaterId);
             if (optionalExaminater.isPresent()) {
                 Examinater examinater = optionalExaminater.get();
